Move more code to AbstractInventory and organize files
