+++ src/main/java/net/minestom/server/inventory/PlayerInventory.java
@@ -10,13 +10,9 @@ import net.minestom.server.event.item.EntityEquipEvent;
 import net.minestom.server.inventory.click.*;
 import net.minestom.server.item.ItemStack;
 import net.minestom.server.item.Material;
-import net.minestom.server.network.packet.server.play.CloseWindowPacket;
-import net.minestom.server.network.packet.server.play.SetSlotPacket;
-import net.minestom.server.network.packet.server.play.WindowItemsPacket;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.List;
 import java.util.stream.IntStream;
 
 /**
@@ -67,62 +63,15 @@ public class PlayerInventory extends AbstractInventory {
             (player, inventory, item, slot) -> IntIterators.fromTo(1, inventory.getSize())
     );
 
-    protected final ClickPreprocessor clickPreprocessor = new ClickPreprocessor(this);
-
-    public PlayerInventory() {
-        super(INVENTORY_SIZE);
+    public static @NotNull IntIterator getInnerShiftClickSlots() {
+        return IntIterators.fromTo(ClickPreprocessor.PLAYER_INVENTORY_OFFSET + 9, ClickPreprocessor.PLAYER_INVENTORY_OFFSET + 45);
     }
 
-    @Override
-    public synchronized void clear() {
-        super.clear();
-
-        // Update equipment
-        for (var player : getViewers()) {
-            player.sendPacketToViewersAndSelf(player.getEquipmentsPacket());
-        }
+    public static @NotNull IntIterator getInnerDoubleClickSlots() {
+        return IntIterators.fromTo(ClickPreprocessor.PLAYER_INVENTORY_OFFSET + 9, ClickPreprocessor.PLAYER_INVENTORY_OFFSET + 45);
     }
 
-    @Override
-    protected void UNSAFE_itemInsert(int slot, @NotNull ItemStack itemStack) {
-        for (var player : getViewers()) {
-            final EquipmentSlot equipmentSlot = fromSlotIndex(slot, player.getHeldSlot());
-            if (equipmentSlot == null) continue;
-
-            EntityEquipEvent entityEquipEvent = new EntityEquipEvent(player, itemStack, equipmentSlot);
-            EventDispatcher.call(entityEquipEvent);
-            itemStack = entityEquipEvent.getEquippedItem();
-        }
-
-        super.UNSAFE_itemInsert(slot, itemStack);
-    }
-
-    @Override
-    public void refreshSlot(int slot, @NotNull ItemStack itemStack) {
-        super.refreshSlot(slot, itemStack);
-        sendPacketToViewers(new SetSlotPacket((byte) 0, 0, (short) slot, itemStack));
-
-        for (var player : getViewers()) {
-            var equipmentSlot = fromSlotIndex(slot, player.getHeldSlot());
-            if (equipmentSlot == null) continue;
-
-            player.syncEquipment(equipmentSlot, itemStack);
-        }
-    }
-
-    @Override
-    public void handleClose(@NotNull Player player) {
-        super.handleClose(player);
-        player.sendPacket(new CloseWindowPacket((byte) 0));
-    }
-
-    @Override
-    public void update(@NotNull Player player) {
-        super.update(player);
-        player.sendPacket(new WindowItemsPacket((byte) 0, 0, List.of(itemStacks), cursorPlayersItem.getOrDefault(player, ItemStack.AIR)));
-    }
-
-    private int getSlotIndex(@NotNull EquipmentSlot slot, int heldSlot) {
+    private static int getSlotIndex(@NotNull EquipmentSlot slot, int heldSlot) {
         return switch (slot) {
             case HELMET, CHESTPLATE, LEGGINGS, BOOTS -> slot.armorSlot();
             case OFF_HAND -> OFFHAND_SLOT;
@@ -130,7 +79,7 @@ public class PlayerInventory extends AbstractInventory {
         };
     }
 
-    private @Nullable EquipmentSlot fromSlotIndex(int slot, int heldSlot) {
+    private static @Nullable EquipmentSlot fromSlotIndex(int slot, int heldSlot) {
         return switch (slot) {
             case HELMET_SLOT -> EquipmentSlot.HELMET;
             case CHESTPLATE_SLOT -> EquipmentSlot.CHESTPLATE;
@@ -139,14 +88,6 @@ public class PlayerInventory extends AbstractInventory {
             case OFFHAND_SLOT -> EquipmentSlot.OFF_HAND;
             default -> slot == (HOTBAR_START + heldSlot) ? EquipmentSlot.MAIN_HAND : null;
         };
-    }
-
-    public @NotNull ItemStack getEquipment(@NotNull EquipmentSlot slot, int heldSlot) {
-        return getItemStack(getSlotIndex(slot, heldSlot));
-    }
-
-    public void setEquipment(@NotNull EquipmentSlot slot, int heldSlot, @NotNull ItemStack newValue) {
-        setItemStack(getSlotIndex(slot, heldSlot), newValue);
     }
 
     private static final int[] EXISTING_ADD_SLOTS = IntStream.concat(
@@ -170,6 +111,63 @@ public class PlayerInventory extends AbstractInventory {
             )
     ).toArray();
 
+    public PlayerInventory() {
+        super(INVENTORY_SIZE);
+    }
+
+    @Override
+    public byte getWindowId() {
+        return 0;
+    }
+
+    @Override
+    public void refreshSlot(int slot, @NotNull ItemStack itemStack) {
+        super.refreshSlot(slot, itemStack);
+
+        for (var player : getViewers()) {
+            var equipmentSlot = fromSlotIndex(slot, player.getHeldSlot());
+            if (equipmentSlot == null) continue;
+
+            player.syncEquipment(equipmentSlot, itemStack);
+        }
+    }
+
+    @Override
+    protected void UNSAFE_itemInsert(int slot, @NotNull ItemStack itemStack) {
+        for (var player : getViewers()) {
+            final EquipmentSlot equipmentSlot = fromSlotIndex(slot, player.getHeldSlot());
+            if (equipmentSlot == null) continue;
+
+            EntityEquipEvent entityEquipEvent = new EntityEquipEvent(player, itemStack, equipmentSlot);
+            EventDispatcher.call(entityEquipEvent);
+            itemStack = entityEquipEvent.getEquippedItem();
+        }
+
+        super.UNSAFE_itemInsert(slot, itemStack);
+    }
+
+    @Override
+    public synchronized void clear() {
+        super.clear();
+
+        for (var player : getViewers()) {
+            player.sendPacketToViewersAndSelf(player.getEquipmentsPacket());
+        }
+    }
+
+    @Override
+    public @Nullable ClickResult handleClick(@NotNull Player player, @NotNull ClickInfo clickInfo) {
+        return CLICK_HANDLER.handleClick(this, player, clickInfo);
+    }
+
+    public @NotNull ItemStack getEquipment(@NotNull EquipmentSlot slot, int heldSlot) {
+        return getItemStack(getSlotIndex(slot, heldSlot));
+    }
+
+    public void setEquipment(@NotNull EquipmentSlot slot, int heldSlot, @NotNull ItemStack newValue) {
+        setItemStack(getSlotIndex(slot, heldSlot), newValue);
+    }
+
     @Override
     public <T> @NotNull T addItemStack(@NotNull ItemStack itemStack, @NotNull TransactionOption<T> option) {
         return processItemStack(itemStack, TransactionType.add(() -> IntIterators.wrap(EXISTING_ADD_SLOTS), () -> IntIterators.wrap(AIR_ADD_SLOTS)), option);
@@ -180,16 +178,4 @@ public class PlayerInventory extends AbstractInventory {
         return processItemStack(itemStack, TransactionType.take(() -> IntIterators.wrap(TAKE_SLOTS)), option);
     }
 
-    @Override
-    public @Nullable ClickResult handleClick(@NotNull Player player, @NotNull ClickInfo clickInfo) {
-        return handleClick(this, CLICK_HANDLER, player, clickInfo);
-    }
-
-    public static @NotNull IntIterator getInnerShiftClickSlots(@NotNull Player player, @NotNull AbstractInventory inventory, @NotNull ItemStack item, int slot) {
-        return IntIterators.fromTo(ClickPreprocessor.PLAYER_INVENTORY_OFFSET + 9, ClickPreprocessor.PLAYER_INVENTORY_OFFSET + 45);
-    }
-
-    public static @NotNull IntIterator getInnerDoubleClickSlots(@NotNull Player player, @NotNull AbstractInventory inventory, @NotNull ItemStack item, int slot) {
-        return IntIterators.fromTo(ClickPreprocessor.PLAYER_INVENTORY_OFFSET + 9, ClickPreprocessor.PLAYER_INVENTORY_OFFSET + 45);
-    }
 }
