+++ src/main/java/net/minestom/server/inventory/click/ClickHandler.java
@@ -1,6 +1,9 @@
 package net.minestom.server.inventory.click;
 
 import net.minestom.server.entity.Player;
+import net.minestom.server.event.EventDispatcher;
+import net.minestom.server.event.inventory.InventoryClickEvent;
+import net.minestom.server.event.inventory.InventoryPreClickEvent;
 import net.minestom.server.inventory.AbstractInventory;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -11,6 +14,50 @@ import org.jetbrains.annotations.Nullable;
  * multiple inventories, but it's also possible to store the inventory yourself and control usages of it.
  */
 public interface ClickHandler {
+
+    /**
+     * Handles the provided click from the given player, returning the results after it is applied. If the results are
+     * null, this indicates that the click was cancelled or was otherwise not processed.
+     *
+     * @param inventory the clicked inventory
+     * @param player the player that clicked
+     * @param clickInfo the information about the player's click
+     * @return the results of the click, or null if the click was cancelled or otherwise was not handled
+     */
+    default @Nullable ClickResult handleClick(@NotNull AbstractInventory inventory, @NotNull Player player, @NotNull ClickInfo clickInfo) {
+        // Call a pre-click event with the base click info
+        var preClickEvent = new InventoryPreClickEvent(player.getInventory(), inventory, player, clickInfo);
+        EventDispatcher.call(preClickEvent);
+        clickInfo = preClickEvent.getClickInfo();
+
+        ClickResult changes = null;
+
+        // Apply the click handler if the click will still occur
+        if (!preClickEvent.isCancelled()) {
+            changes = tryHandle(clickInfo, player, inventory);
+        }
+
+        // Apply each of the conditions to the changes, updating it as we go along
+        for (var condition : inventory.getInventoryConditions()) {
+            changes = condition.accept(player, clickInfo, changes);
+        }
+
+        // Apply the changes and send out an event if there are actually any changes
+        if (changes != null) {
+            changes.applyChanges(player, inventory);
+
+            var clickEvent = new InventoryClickEvent(player, inventory, clickInfo, changes);
+            EventDispatcher.call(clickEvent);
+        }
+
+        // Make sure to update the inventory if indicated
+        if (preClickEvent.shouldUpdate() || changes == null) {
+            preClickEvent.getPlayerInventory().update(player);
+            preClickEvent.getInventory().update(player);
+        }
+
+        return changes;
+    }
 
     /**
      * Handles the provided click info without knowing its type, returning null changes if the type could not be
